+++ backend/services/product-service/src/main/java/vn/edu/uit/msshop/product/variant/domain/model/VariantPrice.java
@@ -3,7 +3,7 @@ package vn.edu.uit.msshop.product.variant.domain.model;
 import vn.edu.uit.msshop.product.shared.domain.exception.DomainException;
 
 public record VariantPrice(
-        int value) {
+        long value) {
     public VariantPrice {
         if (value < 0) {
             throw new DomainException("Variant price must not be negative");
